+++ database/migrations/2023_08_09_084926_create_activities_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('activities', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('term_id');
+            $table->integer('term_id');
             $table->string('title_zh');
             $table->string('title_en')->nullable();
             $table->text('description')->nullable();
